+++ src/main/kotlin/bus/console/models/BusModel.kt
@@ -1,3 +1,3 @@
 package bus.console.models
 
-data class BusModel(var BusID: Int =0, var Route: Int=0, var Origin: String="", var Destination: String="", var Departuretime: Int =0, var arrivaltime: Int =0)
+data class BusModel(var BusID: Int =0, var Route: Int=0, var Origin: String="", var Destination: String="", var Departuretime: Double =0.0, var arrivaltime: Double =0.0)
